+++ Gemfile
@@ -3,5 +3,5 @@ source "https://rubygems.org"
 
 git_source(:github) { |repo_name| "https://github.com/#{repo_name}" }
 
-# Specify your gem's dependencies in github-authentication-provider.gemspec
+# Specify your gem's dependencies in github-authentication.gemspec
 gemspec
