--- lib/github/authentication.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-require "github/authentication/version"
-require "github/authentication/generator"
-require "github/authentication/provider"
-require "github/authentication/cache"
-require "github/authentication/object_cache"
