--- lib/github/authentication/generator.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-require "github/authentication/generator/app"
-require "github/authentication/generator/personal"
