--- lib/github/authentication/version.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module Github
-  module Authentication
-    VERSION = "0.1.2"
-  end
-end
