+++ lib/github_authentication.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+require "github_authentication/version"
+require "github_authentication/generator"
+require "github_authentication/provider"
+require "github_authentication/cache"
+require "github_authentication/object_cache"
+
+module GithubAuthentication
+end
