+++ lib/github_authentication/generator.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+require "github_authentication/generator/app"
+require "github_authentication/generator/personal"
