+++ lib/github_authentication/version.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module GithubAuthentication
+  VERSION = "0.2.0"
+end
