+++ test/test_helper.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 $LOAD_PATH.unshift(File.expand_path("../../lib", __FILE__))
-require "github/authentication/provider"
+require "github_authentication/provider"
 
 require "minitest/autorun"
